refactor: withAmountUtils
