+++ packages/inter-protocol/test/auction/auctionContract.test.js
@@ -38,10 +38,12 @@ import { setUpInstallations } from './tools.js';
 /** @type {import('ava').TestFn<Awaited<ReturnType<makeTestContext>>>} */
 const test = anyTest;
 
+/** @import {AmountUtils} from '@agoric/zoe/tools/test-utils.js'; */
+
 /**
  * @typedef {Record<string, any> & {
- *   bid: IssuerKit & import('../supports.js').AmountUtils;
- *   collateral: IssuerKit & import('../supports.js').AmountUtils;
+ *   bid: IssuerKit & AmountUtils;
+ *   collateral: IssuerKit & AmountUtils;
  *   zoe: ZoeService;
  * }} Context
  */
