+++ packages/inter-protocol/test/supports.js
@@ -1,24 +1,22 @@
-import { AmountMath } from '@agoric/ertp';
 import binaryVoteCounterBundle from '@agoric/governance/bundles/bundle-binaryVoteCounter.js';
 import committeeBundle from '@agoric/governance/bundles/bundle-committee.js';
 import contractGovernorBundle from '@agoric/governance/bundles/bundle-contractGovernor.js';
 import puppetContractGovernorBundle from '@agoric/governance/bundles/bundle-puppetContractGovernor.js';
-import * as utils from '@agoric/vats/src/core/utils.js';
-import { makePromiseSpace, makeAgoricNamesAccess } from '@agoric/vats';
-import { makeFakeBoard } from '@agoric/vats/tools/board-utils.js';
-import { makeMockChainStorageRoot } from '@agoric/internal/src/storage-test-utils.js';
-import { setUpZoeForTest as generalSetUpZoeForTest } from '@agoric/zoe/tools/setup-zoe.js';
-import { makeRatio } from '@agoric/zoe/src/contractSupport/ratio.js';
-import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
-import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
-import { E } from '@endo/far';
 import { makeTracer } from '@agoric/internal';
+import { makeMockChainStorageRoot } from '@agoric/internal/src/storage-test-utils.js';
+import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
 import {
   makeNotifierFromAsyncIterable,
   observeIteration,
   subscribeEach,
 } from '@agoric/notifier';
+import { makeAgoricNamesAccess, makePromiseSpace } from '@agoric/vats';
 import { produceDiagnostics } from '@agoric/vats/src/core/basic-behaviors.js';
+import * as utils from '@agoric/vats/src/core/utils.js';
+import { makeFakeBoard } from '@agoric/vats/tools/board-utils.js';
+import buildManualTimer from '@agoric/zoe/tools/manualTimer.js';
+import { setUpZoeForTest as generalSetUpZoeForTest } from '@agoric/zoe/tools/setup-zoe.js';
+import { E } from '@endo/far';
 
 /** @import {EconomyBootstrapPowers as Space} from '../src/proposals/econ-behaviors.js' */
 
@@ -152,25 +150,7 @@ export const produceInstallations = (space, installations) => {
 
 export const scale6 = x => BigInt(Math.round(x * 1_000_000));
 
-/** @param {Pick<IssuerKit<'nat'>, 'brand' | 'issuer' | 'mint'>} kit */
-export const withAmountUtils = kit => {
-  const decimalPlaces = kit.issuer.getDisplayInfo?.()?.decimalPlaces ?? 6;
-  return {
-    ...kit,
-    /** @param {NatValue} v */
-    make: v => AmountMath.make(kit.brand, v),
-    makeEmpty: () => AmountMath.makeEmpty(kit.brand),
-    /**
-     * @param {NatValue} n
-     * @param {NatValue} [d]
-     */
-    makeRatio: (n, d) => makeRatio(n, kit.brand, d),
-    /** @param {number} n */
-    units: n =>
-      AmountMath.make(kit.brand, BigInt(Math.round(n * 10 ** decimalPlaces))),
-  };
-};
-/** @typedef {ReturnType<typeof withAmountUtils>} AmountUtils */
+export { withAmountUtils } from '@agoric/zoe/tools/test-utils.js';
 
 /** @param {ERef<StoredSubscription<unknown> | StoredSubscriber<unknown>>} subscription */
 export const subscriptionKey = subscription => {
