+++ packages/inter-protocol/test/vaultFactory/driver.js
@@ -33,7 +33,10 @@ import {
   withAmountUtils,
 } from '../supports.js';
 
-/** @import {VaultFactoryContract as VFC} from '../../src/vaultFactory/vaultFactory.js' */
+/**
+ * @import {VaultFactoryContract as VFC} from '../../src/vaultFactory/vaultFactory.js';
+ * @import {AmountUtils} from '@agoric/zoe/tools/test-utils.js';
+ */
 
 const trace = makeTracer('VFDriver');
 
@@ -60,7 +63,7 @@ const contractRoots = {
 /**
  * dL: 1M, lM: 105%, lP: 10%, iR: 100, lF: 500, lP: 0%
  *
- * @param {import('../supports.js').AmountUtils} debt
+ * @param {AmountUtils} debt
  */
 const defaultParamValues = debt =>
   harden({
@@ -78,7 +81,7 @@ const defaultParamValues = debt =>
 
 /**
  * @typedef {{
- *   aeth: IssuerKit & import('../supports.js').AmountUtils;
+ *   aeth: IssuerKit & AmountUtils;
  *   aethInitialLiquidity: Amount<'nat'>;
  *   consume: import('../../src/proposals/econ-behaviors.js').EconomyBootstrapPowers['consume'];
  *   puppetGovernors: {
@@ -95,7 +98,7 @@ const defaultParamValues = debt =>
  *   minInitialDebt: bigint;
  *   reserveCreatorFacet: ERef<AssetReserveLimitedCreatorFacet>;
  *   rates: any;
- *   run: IssuerKit & import('../supports.js').AmountUtils;
+ *   run: IssuerKit & AmountUtils;
  *   stableInitialLiquidity: Amount<'nat'>;
  *   timer: ReturnType<typeof buildManualTimer>;
  *   zoe: ZoeService;
@@ -370,7 +373,7 @@ export const makeManagerDriver = async (
       notification: () => notification,
       /**
        * @param {bigint} collValue
-       * @param {import('../supports.js').AmountUtils} collUtils
+       * @param {AmountUtils} collUtils
        * @param {bigint} [mintedValue]
        */
       giveCollateral: async (collValue, collUtils, mintedValue = 0n) => {
@@ -391,7 +394,7 @@ export const makeManagerDriver = async (
       },
       /**
        * @param {bigint} mintedValue
-       * @param {import('../supports.js').AmountUtils} collUtils
+       * @param {AmountUtils} collUtils
        * @param {bigint} [collValue]
        */
       giveMinted: async (mintedValue, collUtils, collValue = 0n) => {
