+++ packages/inter-protocol/test/vaultFactory/replacePriceAuthority.test.js
@@ -39,12 +39,15 @@ import {
 import '../../src/vaultFactory/types-ambient.js';
 import { defaultParamValues } from './vaultFactoryUtils.js';
 
-/** @import {VaultFactoryContract as VFC} from '../../src/vaultFactory/vaultFactory.js' */
+/**
+ * @import {VaultFactoryContract as VFC} from '../../src/vaultFactory/vaultFactory.js';
+ * @import {AmountUtils} from '@agoric/zoe/tools/test-utils.js';
+ */
 
 /**
  * @typedef {Record<string, any> & {
- *   aeth: IssuerKit & import('../supports.js').AmountUtils;
- *   run: IssuerKit & import('../supports.js').AmountUtils;
+ *   aeth: IssuerKit & AmountUtils;
+ *   run: IssuerKit & AmountUtils;
  *   bundleCache: Awaited<ReturnType<typeof unsafeMakeBundleCache>>;
  *   rates: VaultManagerParamValues;
  *   interestTiming: InterestTiming;
